move back matches from utils to Model
